springboot jpa 动态数据源事务测试，hibernate打印参数

+++ springboot.dynamic.multiple.datasource.jpa/src/main/java/com/example/springboot/study/language/dao/ScStudentEntityRepository.java
@@ -3,5 +3,5 @@ package com.example.springboot.study.language.dao;
 import com.example.springboot.study.language.bean.ScStudentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ScStudentEntityRepository extends JpaRepository<ScStudentEntity, Long> {
+public interface ScStudentEntityRepository extends JpaRepository<ScStudentEntity, Integer> {
 }
